Move pure dtype conversion functions to `dtype.cpp`
