+++ torch_xla/csrc/BUILD
@@ -112,6 +112,7 @@ ptxla_cc_library(
     deps = [
         ":aten_cpu_fallback",
         ":device",
+        ":dtype",
         ":einsum_utilities",
         ":ir",
         ":ir_builder",
@@ -168,6 +169,21 @@ ptxla_cc_library(
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/types:optional",
         "@xla//xla/client:xla_builder",
+    ],
+)
+
+ptxla_cc_library(
+    name = "dtype",
+    srcs = ["dtype.cpp"],
+    hdrs = ["dtype.h"],
+    deps = [
+        "//torch_xla/csrc:device",
+        "//torch_xla/csrc/runtime:tf_logging",
+        "//torch_xla/csrc/runtime:debug_macros",
+        "//torch_xla/csrc/runtime:sys_util",
+        "//torch_xla/csrc/runtime:util",
+        "@com_google_absl//absl/strings",
+        "@com_google_absl//absl/types:optional",
     ],
 )
