+++ torch_xla/csrc/convert_ops.cpp
@@ -3,6 +3,7 @@
 #include <climits>
 
 #include "torch_xla/csrc/aten_xla_bridge.h"
+#include "torch_xla/csrc/dtype.h"
 #include "torch_xla/csrc/helpers.h"
 #include "torch_xla/csrc/runtime/debug_macros.h"
 #include "torch_xla/csrc/tensor_util.h"
@@ -103,7 +104,7 @@ xla::XlaOp ConvertToNumeric(xla::XlaOp op, xla::PrimitiveType from) {
   if (from == xla::PrimitiveType::PRED) {
     torch::lazy::BackendDevice xla_device = bridge::GetCurrentDevice();
     op = ConvertTo(op, from,
-                   GetDevicePrimitiveType(xla::PrimitiveType::U8, &xla_device),
+                   GetDevicePrimitiveType(xla::PrimitiveType::U8, xla_device),
                    &xla_device);
   }
   return op;
