+++ torch_xla/csrc/tensor_util.h
@@ -80,14 +80,9 @@ void PopulateTensorBuffer(const at::Tensor& tensor,
 xla::Shape CreateComputationShapeFromTensor(
     const at::Tensor& tensor, const torch::lazy::BackendDevice* device);
 
-at::ScalarType TensorTypeFromXlaType(xla::PrimitiveType xla_type);
-
-xla::PrimitiveType TensorTypeToRawXlaType(at::ScalarType scalar_type);
-
-// Maps an XLA type to the one which can be used on the given device (or the
-// default device, id device is nullptr).
-xla::PrimitiveType GetDevicePrimitiveType(
-    xla::PrimitiveType type, const torch::lazy::BackendDevice* device);
+// Make a compatible dtype for the current device
+xla::PrimitiveType GetXlaPrimitiveTypeForCurrentDevice(
+    xla::PrimitiveType xla_type);
 
 // Converts the given scalar type to an XLA primitive type.
 xla::PrimitiveType MakeXlaPrimitiveType(
