+++ torch_xla/csrc/xla_lower_util.cpp
@@ -351,8 +351,7 @@ std::vector<xla::XlaOp> CreateKthValue(xla::XlaOp input, int64_t k, int64_t dim,
   }
   // aten::kthvalue() wants Long tensors as indices.
   return {values, xla::ConvertElementType(
-                      indices, GetDevicePrimitiveType(xla::PrimitiveType::S64,
-                                                      /*device=*/nullptr))};
+                      indices, GetXlaPrimitiveTypeForCurrentDevice(xla::PrimitiveType::S64))};
 }
 
 std::vector<xla::XlaOp> CreateTopK(xla::XlaOp input, int64_t k, int64_t dim,
@@ -384,8 +383,7 @@ std::vector<xla::XlaOp> CreateTopK(xla::XlaOp input, int64_t k, int64_t dim,
                                   start_indices, limit_indices, strides);
   // aten::topk() wants Long tensors as indices.
   return {values, xla::ConvertElementType(
-                      indices, GetDevicePrimitiveType(xla::PrimitiveType::S64,
-                                                      /*device=*/nullptr))};
+                      indices, GetXlaPrimitiveTypeForCurrentDevice(xla::PrimitiveType::S64))};
 }
 
 xla::XlaOp CreateMatMul(xla::XlaOp lhs, xla::XlaOp rhs) {
